Use union instead of intersection

See merge request fbedv/radicale/radicale-share-collections!1

+++ radicale_share_collections/__main__.py
@@ -17,30 +17,20 @@ def symlink_shared_collections(storepath, rights, owner, collection, users):
     collection_path = path.join(owner, collection)
     collection_dir = path.join(storepath, collection_path)
     for user in users.difference({owner}):
-        has_read = 'r' in rights.authorization(user, "/" + collection_path)
-        destination = path.join(
-            storepath, user, "from" + "-" + owner + "-" + collection
-        )
-        if has_read:
-            if path.exists(destination):
-                continue
-            else:
-                symlink(collection_dir, destination)
-        else:
-            if path.islink(destination):
-                unlink(destination)
+        has_read = 'r' in rights.authorization(user, f"/{collection_path}")
+        destination = path.join(storepath, user, f"from-{owner}-{collection}")
+        if has_read and not path.exists(destination):
+            symlink(collection_dir, destination)
+        elif not has_read and path.islink(destination):
+            unlink(destination)
 
 
 def delete_broken_symlinks(collection_path):
-    links = [
-        path.join(collection_path, link)
-        for link in listdir(collection_path)
-        if path.islink(path.join(collection_path, link))
-    ]
-    for link in links:
-        # check if symlink is broken then unlink it
-        if path.lexists(link) and not path.exists(link):
-            unlink(link)
+    # find broken symlinks and unlink them
+    with scandir(collection_path) as dirit:
+        for link in dirit:
+            if link.is_symlink() and not path.exists(link):
+                unlink(link.path)
 
 
 def manage_symlinks(storepath, rights, collections, users):
@@ -88,13 +78,13 @@ all collections.""",
     if args.users is None:
         users = primary_collections
     else:
-        users = set(args.users.strip(",").split(",")).union(
+        users = set(args.users.strip(",").split(",")).intersection(
             primary_collections
         )
     if args.collections is None:
         collections = primary_collections
     else:
-        collections = set(args.collections.strip(",").split(",")).union(
+        collections = set(args.collections.strip(",").split(",")).intersection(
             primary_collections
         )
     manage_symlinks(storepath, rights, collections, users)
